Tweak menu positioning
