+++ src/menu/containers/Toolbar.tsx
@@ -8,14 +8,13 @@ import {
   redo,
   toggleAutoSave,
 } from '../../editor/editorSlice'
-import { toggleMenu, closeMenu, setShouldRender } from '../menuSlice'
+
 import {
   useEditorText,
   useFutureLength,
   usePastLength,
   useAutoSaveEnabled,
 } from '../../editor/hooks'
-import { useIsMenuOpen, useShouldRenderMenu } from '../hooks'
 
 import ToolbarComponent from '../components/Toolbar'
 
@@ -24,9 +23,6 @@ const Toolbar: React.FC = () => {
   const pastLength = usePastLength()
   const futureLength = useFutureLength()
   const autoSaveEnabled = useAutoSaveEnabled()
-
-  const isMenuOpen = useIsMenuOpen()
-  const shouldRenderMenu = useShouldRenderMenu()
 
   const dispatch = useDispatch()
 
@@ -52,21 +48,6 @@ const Toolbar: React.FC = () => {
     [dispatch, autoSaveEnabled],
   )
 
-  const handleToggleMenu = useCallback(() => {
-    dispatch(toggleMenu())
-  }, [dispatch])
-
-  const setShouldRenderMenu = useCallback(
-    (shouldRender: boolean) => {
-      dispatch(setShouldRender(shouldRender))
-    },
-    [dispatch],
-  )
-
-  const handleCloseMenu = useCallback(() => {
-    dispatch(closeMenu())
-  }, [dispatch])
-
   return (
     <ToolbarComponent
       content={content}
@@ -78,11 +59,6 @@ const Toolbar: React.FC = () => {
       onOpen={handleOpen}
       onUndo={handleUndo}
       onRedo={handleRedo}
-      isOpen={isMenuOpen}
-      shouldRender={shouldRenderMenu}
-      toggleMenu={handleToggleMenu}
-      closeMenu={handleCloseMenu}
-      setShouldRender={setShouldRenderMenu}
     />
   )
 }
